Correctly import default get from lodash.get (#68)
* Update docs to reflect v4 and add testing block. (#60)

* Drop normalizer version back and fix jest test

* Bump node version and fix Object.values in 6

* Add Node version to Requirements section.

* Correctly import default get from lodash.get

`lodash.get` exports the get function as its default export. Importing it as though it were an individually exported module (rather than the default) doesn't work–it results in an undefined function error.

* Correctly import default get from lodash.get (#1)

`lodash.get` exports the get function as its default export. Importing it as though it were an individually exported module (rather than the default) doesn't work–it results in an undefined function error.

* V4.2.0 (#2)

* docs for draft posts

* support dot-paths for shouldUpdate

* rename AckRequest -> RequestAck

* update jsdocs

* use castarray

* move non-invariant out of invariants.js

* refactor query counter check

* update deps and tests

* rem readme feature

* preserve query result functionality

* v4.2.0-beta.0

* Add support for query generators (#59)

* Add support for query generators

* Update docs to reflect v4 and add testing block. (#60)

* version bump

* code style nitpicks

* Correctly import default get from lodash.get (#1)

`lodash.get` exports the get function as its default export. Importing it as though it were an individually exported module (rather than the default) doesn't work–it results in an undefined function error.